file renames and some fixes for PublicOrEmbeddedDashboard
