--- frontend/src/metabase/public/containers/PublicDashboard/index.ts
@@ -1,1 +1,0 @@
-export { PublicOrEmbeddedDashboardControlled } from "./PublicDashboard";
