+++ frontend/src/metabase/public/containers/PublicOrEmbeddedDashboard/PublicOrEmbeddedDashboard.unit.spec.tsx
@@ -8,13 +8,13 @@ import {
 } from "metabase-types/api/mocks";
 import { createMockState } from "metabase-types/store/mocks";
 
-import { PublicOrEmbeddedDashboardControlled } from "./PublicDashboard";
+import { PublicOrEmbeddedDashboardControlled } from "./PublicOrEmbeddedDashboard";
 
 const MOCK_TOKEN =
   "eyJhbGciOiJIUzI1NiJ9.eyJyZXNvdXJjZSI6eyJkYXNoYm9hcmQiOjExfSwicGFyYW1zIjp7fSwiaWF0IjoxNzEyNjg0NTA1LCJfZW1iZWRkaW5nX3BhcmFtcyI6e319.WbZTB-cQYh4gjh61ZzoLOcFbJ6j6RlOY3GS4fwzv3W4";
 const DASHBOARD_TITLE = '"My test dash"';
 
-describe("PublicDashboard", () => {
+describe("PublicOrEmbeddedDashboard", () => {
   it("should display dashboard tabs", async () => {
     await setup({ numberOfTabs: 2 });
