+++ frontend/src/metabase/public/containers/PublicOrEmbeddedDashboard/WithPublicDashboardEndpoints.tsx
