+++ frontend/src/metabase/public/containers/PublicOrEmbeddedDashboard/index.ts
@@ -1,0 +1,1 @@
+export { PublicOrEmbeddedDashboardControlled } from "./PublicOrEmbeddedDashboard";
