+++ frontend/src/metabase/routes-embed.jsx
@@ -2,7 +2,7 @@ import { Route } from "react-router";
 
 import { PublicNotFound } from "metabase/public/components/PublicNotFound";
 import PublicApp from "metabase/public/containers/PublicApp";
-import { PublicOrEmbeddedDashboardControlled } from "metabase/public/containers/PublicDashboard";
+import { PublicOrEmbeddedDashboardControlled } from "metabase/public/containers/PublicOrEmbeddedDashboard";
 import { PublicQuestion } from "metabase/public/containers/PublicQuestion";
 
 export const getRoutes = store => (
